Begin the work on the projects HTML and css
Current commit introduces various new files and edits to these files in
order to get the project closer to the end result. For the moment,
the HTML is done and no changes are being added to this currently.
In the CSS file, color was changed to the left area text to match
with the images provided. Flex was introduced to help center and
position the different containers that will be styled. A pseudo "Terms
and Services" page was added to be edited later with the hyperlink text
"Terms and Services" under the "CLAIM YOUR FREE TRIAL" button.