Apply more styles to project
More styles implemented in order to bring the project to the final
result. Working on getting things positioned in their proper areas
such as the small banner/notice at over the form. The spacing between
the lines from the words from the left side text. As well as spacing
the left side text from the form. Working out the proper size of the
form before stylizing it's inner contents.